update: society page
