+++ src/pages/society/Society.css
@@ -1,46 +1,43 @@
-body{
+body {
     background-color: #0A0D14;
     color: #ffffff;
 }
-.profile{
+
+.profile {
     width: 80%;
     display: flex;
     padding: 2rem 0rem;
     margin: 4rem 6rem;
-    justify-content: space-evenly ;
+    justify-content: space-evenly;
     flex-wrap: wrap;
 }
-.profile-left{
+
+.profile-left {
     width: 30%;
-    /* position: absolute; */
-    /* left: 10%; */
-   
-    
 }
-.profile-right{
+
+.profile-logo {
+    height: 15rem;
+    width: 90%;
+    object-fit: contain;
+}
+
+.profile-right {
     width: 70%;
     display: flex;
     flex-direction: column;
-    /* position: relative;
-    left: 10%; */
-}
-.profile-right p{
-    font-size: 1.25rem;
-}
-.profile-logo{
-    height: 15rem;
-    /* border-radius: 50%; */
+    padding-left: 2rem;
 }
 
-.society h2{
-    /* position: relative; */
-    font-size: 3rem;
-    color: #efcb44;
-    /* font-family: ; */
-    /* -webkit-text-stroke: 0.3vw #EFCA46; */
-    /* color: #EFCA46; */
-    /* text-transform: uppercase; */
+.profile-right p {
+    font-size: 1.25rem;
 }
+
+.society-container h2 {
+    font-size: 2.5rem;
+    color: #efcb44;
+}
+
 /* .society h1::before{
     content: attr(data-text);
     position: absolute;
@@ -56,25 +53,22 @@ body{
     -webkit-animation: animate 6s linear infinite; 
 } */
 
-
-h4{
-    color: #ffffff;
-}
-.about-us{
+.about-us {
     width: 80%;
     margin: 4rem 6rem;
     display: flex;
     flex-direction: column;
 }
-.about-us p{
+
+.about-us p {
     font-size: 1.25rem;
 }
 
-.society_recent_event_section{
+.society_recent_event_section {
     margin: 4rem 6rem;
 }
 
-.society_recentEvent_card{
+.society_recentEvent_card {
     display: flex;
     width: 80%;
     padding: 1.5rem 0;
@@ -82,7 +76,7 @@ h4{
     border-bottom: 0.5px solid white;
 }
 
-.society_recentEvent_card img{
+.society_recentEvent_card img {
     height: 10rem;
     width: 10rem;
     object-fit: cover;
